rename fuzz_html_helpers.py to html_helpers.py
