+++ post-processing/analyses/filepath_analyser.py
@@ -23,7 +23,7 @@ from typing import (
 )
 
 import fuzz_analysis
-import fuzz_html_helpers
+import html_helpers
 
 import datatypes.project_profile
 import datatypes.fuzzer_profile
@@ -66,7 +66,7 @@ class FuzzFilepathAnalyser(fuzz_analysis.AnalysisInterface):
         html_string += "<div class=\"report-box\">"
 
         # Table with all files
-        html_string += fuzz_html_helpers.html_add_header_with_link(
+        html_string += html_helpers.html_add_header_with_link(
             "Files and Directories in report",
             1,
             toc_list
@@ -85,13 +85,13 @@ class FuzzFilepathAnalyser(fuzz_analysis.AnalysisInterface):
             "Config.md#code-exclusion-from-the-report\">link</a></p>"
         )
 
-        html_string += fuzz_html_helpers.html_add_header_with_link(
+        html_string += html_helpers.html_add_header_with_link(
             "Files in report",
             2,
             toc_list
         )
         tables.append(f"myTable{len(tables)}")
-        html_string += fuzz_html_helpers.html_create_table_head(
+        html_string += html_helpers.html_create_table_head(
             tables[-1],
             [
                 ("Source file", ""),
@@ -114,7 +114,7 @@ class FuzzFilepathAnalyser(fuzz_analysis.AnalysisInterface):
                 if is_file_covered:
                     profiles_that_cover.append(profile.get_key())
 
-            html_string += fuzz_html_helpers.html_table_add_row(
+            html_string += html_helpers.html_table_add_row(
                 [
                     f"{fnm}",
                     f"{str(profiles_that_hit)}",
@@ -124,20 +124,20 @@ class FuzzFilepathAnalyser(fuzz_analysis.AnalysisInterface):
         html_string += "</table>"
 
         # Table with all directories
-        html_string += fuzz_html_helpers.html_add_header_with_link(
+        html_string += html_helpers.html_add_header_with_link(
             "Directories in report",
             2,
             toc_list
         )
         tables.append(f"myTable{len(tables)}")
-        html_string += fuzz_html_helpers.html_create_table_head(
+        html_string += html_helpers.html_create_table_head(
             tables[-1],
             [
                 ("Directory", ""),
             ]
         )
         for dr in all_proj_dirs:
-            html_string += fuzz_html_helpers.html_table_add_row([f"{dr}"])
+            html_string += html_helpers.html_table_add_row([f"{dr}"])
         html_string += "</table>"
 
         html_string += "</div>"  # .collapsible
